Ramp.MNT: organize subpackage and improve file names
